requirements.txt
crewai==0.108.0
crewai_tools==0.38.1
fitz==0.0.1.dev2
pdfminer_six==20250327
pydantic==2.11.1
python-dotenv==1.1.0
python_docx==1.1.2
